experimentation with spring data jdbc
