+++ infobip-spring-data-jdbc-querydsl/src/main/java/com/infobip/spring/data/jdbc/EnableQuerydslJdbcRepositories.java
@@ -1,0 +1,16 @@
+package com.infobip.spring.data.jdbc;
+
+import org.springframework.data.jdbc.repository.config.EnableJdbcRepositories;
+
+import java.lang.annotation.*;
+
+/**
+ * Annotation to enable {@link QuerydslJdbcRepository} support.
+ *
+ * @see EnableJdbcRepositories
+ */
+@EnableJdbcRepositories(repositoryFactoryBeanClass = QuerydslJdbcRepositoryFactoryBean.class)
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface EnableQuerydslJdbcRepositories {
+}
