+++ infobip-spring-data-jdbc-querydsl/src/test/java/com/infobip/spring/data/jdbc/PersonSettings.java
@@ -1,0 +1,14 @@
+package com.infobip.spring.data.jdbc;
+
+import lombok.Value;
+import lombok.With;
+import org.springframework.data.annotation.Id;
+
+@Value
+class PersonSettings {
+
+	@With
+	@Id
+	private final Long id;
+	private final Long personId;
+}
