+++ infobip-spring-data-jdbc-querydsl/src/test/resources/db/migration/V1_0_0__base_schema.sql
