+++ infobip-spring-data-jpa-querydsl/src/main/java/com/infobip/spring/data/jpa/EnableExtendedJpaRepositories.java
@@ -1,0 +1,16 @@
+package com.infobip.spring.data.jpa;
+
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+import java.lang.annotation.*;
+
+/**
+ * Annotation to enable {@link ExtendedQuerydslJpaRepository} support.
+ *
+ * @see EnableJpaRepositories
+ */
+@EnableJpaRepositories(repositoryBaseClass = SimpleExtendedQuerydslJpaRepository.class)
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface EnableExtendedJpaRepositories {
+}
