+++ infobip-spring-data-jpa-querydsl/src/test/java/com/infobip/spring/data/jpa/Person.java
@@ -1,0 +1,28 @@
+package com.infobip.spring.data.jpa;
+
+import lombok.*;
+
+import javax.persistence.*;
+
+@Getter
+@ToString
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@Entity
+class Person {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column
+	private Long id;
+
+	@Column
+	private String firstName;
+
+	@Column
+	private String lastName;
+
+	Person(String firstName, String lastName) {
+		this.firstName = firstName;
+		this.lastName = lastName;
+	}
+}
