+++ infobip-spring-data-jpa-querydsl/src/test/java/com/infobip/spring/data/jpa/PersonRepository.java
@@ -1,0 +1,4 @@
+package com.infobip.spring.data.jpa;
+
+interface PersonRepository extends ExtendedQuerydslJpaRepository<Person, Long> {
+}
