+++ infobip-spring-data-jpa-querydsl/src/test/java/com/infobip/spring/data/jpa/PersonSettingsRepository.java
@@ -1,0 +1,5 @@
+package com.infobip.spring.data.jpa;
+
+interface PersonSettingsRepository extends ExtendedQuerydslJpaRepository<PersonSettings, Long> {
+
+}
