+++ infobip-spring-data-jpa-querydsl/src/test/resources/application.yaml
