--- src/test/java/com/infobip/spring/data/Person.java
@@ -1,28 +1,0 @@
-package com.infobip.spring.data;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-@Getter
-@ToString
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@Entity
-class Person {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column
-	private Long id;
-
-	@Column
-	private String firstName;
-
-	@Column
-	private String lastName;
-
-	Person(String firstName, String lastName) {
-		this.firstName = firstName;
-		this.lastName = lastName;
-	}
-}
