--- src/test/java/com/infobip/spring/data/PersonRepository.java
@@ -1,4 +1,0 @@
-package com.infobip.spring.data;
-
-interface PersonRepository extends ExtendedQueryDslJpaRepository<Person, Long> {
-}
