--- src/test/java/com/infobip/spring/data/PersonSettings.java
@@ -1,24 +1,0 @@
-package com.infobip.spring.data;
-
-import lombok.*;
-
-import javax.persistence.*;
-
-@Getter
-@ToString
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@Entity
-class PersonSettings {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column
-	private Long id;
-
-	@Column
-	private Long personId;
-
-	PersonSettings(Long personId) {
-		this.personId = personId;
-	}
-}
