--- src/test/java/com/infobip/spring/data/PersonSettingsRepository.java
@@ -1,5 +1,0 @@
-package com.infobip.spring.data;
-
-interface PersonSettingsRepository extends ExtendedQueryDslJpaRepository<PersonSettings, Long> {
-
-}
